% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight.R
\name{as_highlighter}
\alias{as_highlighter}
\title{Convert a highlighted vector into a highlighter}
\usage{
as_highlighter(.x)
}
\description{
Convert a highlighted vector \code{x} (class \code{vlightr_highlight}) into a highlighter
function (class \code{vlightr_highligher}) which applies the same conditional
formatting as that of \code{x}.
}
