% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{attr-setters}
\alias{attr-setters}
\alias{set_conditions}
\alias{set_formatters}
\alias{set_description}
\alias{set_precedence}
\alias{set_format_once}
\alias{set_init_formatter}
\alias{set_last_formatter}
\title{Set an attribute of a highlighted vector}
\usage{
set_conditions(x, value, at = NULL)

set_formatters(x, value, at = NULL)

set_description(x, value, at = NULL)

set_precedence(x, value, at = NULL)

set_format_once(x, value)

set_init_formatter(x, value)

set_last_formatter(x, value)
}
\arguments{
\item{x}{\verb{[vlightr_highlight]}

A highlighted vector to set an attribute of.}

\item{value}{In \verb{set_*}, the attribute \code{*} that you wish to set.

For example, in \code{set_conditions()}, the \code{value} must be a function,
purrr-style lambda, or list of functions or lambdas to be used as the
\code{conditions} of \code{x}.

See \code{\link[=highlight]{highlight()}} for the specifics of each attribute.}

\item{at}{\verb{[logical / numeric]}

For \code{set_conditions()}, \code{set_formatters()}, \code{set_description()}, and
\code{set_precedence()}, the indices at which to set the new attribute. Must
be a logical or numeric vector.}
}
\value{
The vector \code{x} with one of it's attributes updated.
}
\description{
Overwrites the \code{conditions}, \code{formatters}, \code{description}, \code{precedence},
\code{format_once}, \code{init_formatter}, or \code{last_formatter} attribute of an
existing \code{vlighter_highlight} vector and validates the result.
}
\examples{
# Let's make these affirmations into exclamations
affirmations <- c("COOL", "Good stuff", "WOW", "Way to go")
exclamations <- highlight(
  affirmations,
  conditions = ~ toupper(.x) == .x,
  formatters = ~ paste(.x, "!")
)
print(affirmations)
print(exclamations)
describe_highlight(exclamations)

# Reset `formatters` to quiet down the exclamation instead
exclamations <- set_formatters(exclamations, ~ tolower(.x))
print(exclamations)

# Add a description
exclamations <- set_description(exclamations, "Set Uppercase to Lower")
describe_highlight(exclamations)

# The `at` argument is useful with multiple formatters
x_multi <- highlight(
  c(13, 4, 6, 1),
  conditions = list(
    ~ .x \%\% 2 == 0,
    ~ .x \%\% 2 != 0,
    ~ .x > 10
   ),
  formatters = list(
    ~ paste(.x, "is even"),
    ~ paste(.x, "is odd"),
    ~ paste(.x, "is big and")
 ),
 last_formatter = wrap("[", "]")
)
print(x_multi)

# Move the `paste(.x, "is big and")` formatter first
x_multi <- set_precedence(x_multi, value = -1, at = 3)
print(x_multi)
}
\seealso{
\code{\link[=update_highlight]{update_highlight()}} to reset multiple attributes of a highlighted vector at
once.
}
\concept{attribute setters}
