% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition.R
\name{condition_keep}
\alias{condition_keep}
\title{Subset elements of a highlighted vector which meet a condition(s)}
\usage{
condition_keep(x, which = TRUE, meets = c("all", "any", "none"))
}
\arguments{
\item{x}{\verb{[vlightr_highlight]}

A highlighted vector to subset.}

\item{which}{\verb{[character / logical / numeric]}

Which conditions are used to subset the vector. Conditions can be selected by
index using a numeric or logical vector, or selected using the name of a
corresponding description(s) (\verb{attr(x, 'description)}).}

\item{meets}{\verb{[character(1)]}

One of \code{"all"}, \code{"any"}, or \code{"none"}. If \code{"all"}, then elements of \code{x} which
meet all of the specified conditions are subset, if \code{"any"}, then elements
which meet \code{"any"} of the condition are subset, and if \code{"none"}, then elements
which meet \code{"none"} of the conditions are subset}
}
\description{
\code{\link[=condition_keep]{condition_keep()}} keeps elements of a highlighted vector \code{x} which meet the
specified conditions (i.e. receive the same conditional formatting).

\code{condition_keep(x, 1)} is equivalent to \code{x[attr(x, 'conditions')[[1]](x)]}.
}
\examples{
# Define a highlight
summary_hl <- highlight_case(
  x = 1:5,
  .x == min(.x) ~ paste0("(Min = ", .x, ")"),
  .x == quantile(.x, 0.5) ~ paste0("(Mid = ", .x, ")"),
  .x == max(.x) ~ paste0("(Max = ", .x, ")"),
  .x > quantile(.x, 0.5) ~ paste("[ High", .x, "]"),
  .x < quantile(.x, 0.5) ~ paste("[ Low", .x, "]"),
  .description = c(
    "Min",
    "Middle",
    "Max",
    ">P50",
    "<P50"
  )
)

# Keep the middle element
condition_keep(x, which = 3)

# Specify a condition by it's description
condition_keep(x, which = "Middle")

# Keep the minimum and maximum elements
condition_keep(x, which = c("Min", "Max"), meets = "any")
condition_keep(x, which = c("Min", "Max"), meets = "none")
condition_keep(x, which = c("Min", "Max"), meets = "all")
}
\seealso{
Other condition utilities: 
\code{\link{condition_id}()},
\code{\link{condition_met}()},
\code{\link{condition_split}()}
}
\concept{condition utilities}
