% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{formatters}
\alias{formatters}
\alias{formatters<-}
\title{Get or set the formatter functions of a highlight or highlighter}
\usage{
formatters(x)

formatters(x) <- value
}
\description{
Get or set the formatter functions of a highlight or highlighter
}
