% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{highlight_functions}
\alias{highlight_functions}
\alias{highlight_functions<-}
\title{Get or set the test and formatter functions of a highlight or highlighter}
\usage{
highlight_functions(x, as = c("functions", "highlighters"))

highlight_functions(x) <- value
}
\description{
Get or set the test and formatter functions of a highlight or highlighter
}
