% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{highlight_lambda}
\alias{highlight_lambda}
\title{Convert a one-sided formula to a function}
\usage{
highlight_lambda(x, x_name = rlang::caller_arg(x))
}
\description{
Convert a one-sided formula to a function
}
