% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight.R
\name{is_highlight}
\alias{is_highlight}
\title{Test if the object is a highlight}
\usage{
is_highlight(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
\code{TRUE} if \code{x} is a \code{vlightr_highlight}, \code{FALSE} otherwise.
}
\description{
This function returns \code{TRUE} for highlighted vectors (class \code{vlightr_highlight})
or subclasses thereof, and returns returns \code{FALSE} otherwise.
}
\examples{
# The following are `FALSE`
is_highlight(10)

# The following are `TRUE`
is_highlight(highlight(10))
is_highlight(templight(10))
is_highlight(highlight_case(10))

}
