% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight.R
\name{is_highlight_case}
\alias{is_highlight_case}
\title{Test if the object is a highlight-case vector}
\usage{
is_highlight_case(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
\code{TRUE} if \code{x} is a \code{vlightr_highlight_case}, \code{FALSE} otherwise.
}
\description{
This function returns \code{TRUE} for vectors highlighted using \code{highlight_case()}
or \code{templight_case()} (class \code{vlightr_highlight_case}), and returns
\code{FALSE} otherwise.
}
\examples{
# The following are `FALSE`
is_highlight_case(10)
is_highlight_case(highlight(10))

# The following are `TRUE`
is_highlight_case(highlight_case(10))
is_highlight_case(templight_case(10))

# Vectors highlighted using `highlighter_case()` or
# `templigher_case()` highlighter are class `vlightr_highlight_case`
lighter <- highlighter_case()
is_highlight_case(lighter(10))

}
