% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight.R
\name{is_highlighter}
\alias{is_highlighter}
\title{Test if the object is a highlighter}
\usage{
is_highlighter(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
\code{TRUE} if \code{x} is a \code{vlightr_highlighter}, \code{FALSE} otherwise.
}
\description{
This function returns \code{TRUE} for highlighter functions (class \code{vlightr_highlighter})
or subclasses thereof, and returns returns \code{FALSE} otherwise.
}
