% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templight.R
\name{is_templight}
\alias{is_templight}
\title{Test if the object is a temp-light vector}
\usage{
is_templight(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
\code{TRUE} if \code{x} is a \code{vlightr_templight}, \code{FALSE} otherwise.
}
\description{
This function returns \code{TRUE} for temp-lighted vectors (class \code{vlightr_templight})
or subclasses thereof, and returns returns \code{FALSE} otherwise.
}
\examples{
# The following are `FALSE`
is_templight(12)
is_templight(highlight(12))

# The following are `TRUE`
is_templight(templight(12))
is_templight(templight_case(12))

}
