% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templight.R
\name{is_templight_case}
\alias{is_templight_case}
\title{Test if the object is a temp-light-case vector}
\usage{
is_templight_case(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
\code{TRUE} if \code{x} is a \code{vlightr_templight/vlightr_highlight_case}, \code{FALSE} otherwise.
}
\description{
This function returns \code{TRUE} for vectors highlighted using \code{templight_case()}
(class \code{vlightr_templight/vlightr_highlight_case}), and returns \code{FALSE}
otherwise.
}
\examples{
# The following are `FALSE`
is_templight_case(10)
is_templight_case(highlight(10))
is_templight_case(highlight_case(10))

# The following are `TRUE`
is_templight_case(templight_case(10))

# Vectors highlighted using the `templigher_case()` or highlighter are
class `vlightr_templight/vlightr_highlight_case`.
templighter <- templighter_case()
is_templight_case(templighter(10))

}
