% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templight.R
\name{is_templighter}
\alias{is_templighter}
\title{Test if the object is a temp-lighter function}
\usage{
is_templighter(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
\code{TRUE} if \code{x} is a \code{vlightr_templighter}, \code{FALSE} otherwise.
}
\description{
This function returns \code{TRUE} for functions created using \code{templighter()}
or \code{templighter_case()} (class \code{vlightr_templighter}), and returns \code{FALSE}
otherwise.
}
