% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templight.R
\name{is_templighter_case}
\alias{is_templighter_case}
\title{Test if the object is a temp-lighter function}
\usage{
is_templighter_case(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
\code{TRUE} if \code{x} is a \code{vlightr_templighter/vlightr_highlighter_case}, \code{FALSE} otherwise.
}
\description{
This function returns \code{TRUE} for functions created using \code{templighter_case()}
(class \code{vlightr_templighter/vlightr_highlighter_case}), and returns \code{FALSE}
otherwise.
}
