% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{is_vlightr_case}
\alias{is_vlightr_case}
\title{Test if the object is a highlight-case or temp-light-case vector or a
highlighter-case or temp-lighter-case function}
\usage{
is_vlightr_case(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
\code{TRUE} if \code{x} is a {vlightr} "case" object, \code{FALSE} otherwise.
}
\description{
This function returns \code{TRUE} for objects inheriting the following classes:
\itemize{
\item \code{"vlightr_highlight_case"}
\item \code{"vlightr_highlighter_case"}
}

These includes objects created by \code{templight_case()} and \code{templighter_case()}.

This function returns \code{FALSE} if \code{x} is any other object.
}
