% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{label}
\alias{label}
\title{Generate a formatter function which labels a vector}
\usage{
label(string, left = "[", right = "]")
}
\arguments{
\item{string}{\verb{[character(1)]}

The word used to label the vector.}

\item{left, right}{\verb{[character(1)]}

A prefix (\code{left}) or suffix (\code{right}) that the output function pastes around
the label's \code{string}. By default, \code{left} is \code{"["} and \code{right} is \code{"]"}.}
}
\value{
A function that pastes a label onto it's input.
}
\description{
\code{label()} generates a function that pastes a label onto the end of every
element of it's input vector using \code{\link[=paste0]{paste0()}}.
}
\examples{
high <- label("high")
low <- label("low")
unsure <- label("?", "<<", ">>")

high(5)
low(1)
unsure(c(7, -9, NA))
unsure(letters[1:5])

}
