% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{stylers}
\alias{stylers}
\alias{bg}
\alias{background}
\alias{color}
\alias{colour}
\alias{make_formatter}
\alias{style}
\alias{wrap}
\alias{emph}
\title{Generate a formatter function to style text}
\usage{
bg(x)

background(x)

color(x)

colour(x)

make_formatter(x)

style(x)

wrap(left = "[", right = "]")

emph(wrap = TRUE, bg = TRUE, bold = TRUE)
}
\arguments{
\item{x}{\verb{[character(1)]}

For \code{bg()} and \code{color()}.}

\item{left, right}{\verb{[character(1)]}

For \code{wrap()}.}
}
\value{
A function which returns a styled character or ANSI string vector.
}
\description{
vlightr provides some helpers to quickly generate new formatting functions
for use in the \code{formatters}, \code{init_formatter}, and \code{last_formatter} arguments
of \code{\link[=highlight]{highlight()}}, \code{\link[=highlighter]{highlighter()}}, and friends.
}
\examples{
color_red <- color("red")
color_red(1:3)
}
\concept{styles}
