% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{templight_functions}
\alias{templight_functions}
\alias{templight_functions<-}
\alias{locations}
\title{Get or set the locations and formatter functions of a temp-lighted vector or
temp-lighter function.}
\usage{
templight_functions(x, as_list_of = c("list", "templighter"))

templight_functions(x) <- value

locations(x)

locations(x)
}
\description{
Get or set the test and formatter functions of a temp-light or temp-lighter
}
