% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templight.R
\name{templighter}
\alias{templighter}
\alias{templighter_mult}
\alias{templighter_case}
\title{Make a temporary highlighter}
\usage{
templighter(.t, .at)

templighter_mult(...)

templighter_case(...)
}
\description{
Makes a temporary highlighter.
}
