% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{tests}
\alias{tests}
\alias{tests<-}
\title{Get or set the test functions of a highlight or highlighter}
\usage{
tests(x)

tests(x) <- value
}
\description{
Get or set the test functions of a highlight or highlighter
}
