% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{true}
\alias{true}
\title{Return the value \code{TRUE}}
\usage{
true(x)
}
\arguments{
\item{x}{Any object.}
}
\value{
The value \code{TRUE}.
}
\description{
This function always returns the value \code{TRUE}.
}
\examples{
# `true()` is useful for supplying a default formatter
# to `highlight_case()`.
levels <- highlight_case(
  1:5,
  1 ~ "Low",
  2 ~ "Mid",
  3 ~ "High",
  true ~ label("?") # Label any value not in 1, 2, 3
)
print(levels)

}
