% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight.R
\name{un_highlight}
\alias{un_highlight}
\alias{ul}
\title{Unclass a highlighted vector}
\usage{
un_highlight(x)

ul(x)
}
\arguments{
\item{x}{A highlighted vector to unclass. If \code{x} is not a \code{vlightr_highlight} vector,
then \code{x} is returned as is.}
}
\description{
Remove the conditional formatting from a highlighted vector,
returning the highlight to it's underlying class. If \code{x} is a highlightable
vector, then \code{x} and \code{un_highlight(highlight(x))} are equivalent.

This is useful to expose a highlighted vector's data to function which do not
accept a \code{vlightr_highlight} class object.

\code{ul()} and \code{un_highlight()} are synonyms.
}
\examples{
# Remove the conditional formatting of a highlighted vector
x <- highlight(1:5, ~ .x > 3, wrap())
print(x)
print(un_highlight(x))

# Allow a highlighted vector to be used in functions
try(paste(x))
paste(un_highlight(x))
}
