% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vlightr-package.R
\docType{package}
\name{vlightr-package}
\alias{vlightr}
\alias{vlightr-package}
\title{vlightr: Persistent Conditional Formatting for Vectors}
\description{
Provides a generic vector class with persistant conditional formatting. `highlight` an arbitrary vector to color, emphasize, or otherwise reformat elements of interest when printed as a vector or tibble column. Make exploratory data analysis and interactive data validation easier on the eyes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/EthanSansom/vlightr}
  \item \url{https://ethansansom.github.io/vlightr/}
  \item Report bugs at \url{https://github.com/EthanSansom/vlightr/issues}
}

}
\author{
\strong{Maintainer}: Ethan Sansom \email{ethan.sansom29@gmail.com}

}
\keyword{internal}
