% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{wrap}
\alias{wrap}
\title{Generate a formatter function which prefixes and suffixes a vector}
\usage{
wrap(left = "[", right = "]")
}
\arguments{
\item{left, right}{\verb{[character(1)]}

A prefix (\code{left}) or suffix (\code{right}) that the output function pastes onto a
character vector. By default, \code{left} is \code{"["} and \code{right} is \code{"]"}.}
}
\value{
A function that pastes a prefix and suffix onto it's input.
}
\description{
\code{wrap()} generates a function that pastes a prefix \code{left} and a suffix \code{right}
onto every element of an input vector using \code{\link[=paste0]{paste0()}}.
}
\examples{
default <- wrap()
pointy <- wrap("<", ">")
question <- wrap("", "?")

default(1:5)
pointy(c("a", "b", "c"))
question(c("who", "what", "where", "why"))

}
