% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition.R
\name{condition_id}
\alias{condition_id}
\title{Identify elements of a highlighted vector which meet a condition(s)}
\usage{
condition_id(x, no_id = NA_character_)
}
\arguments{
\item{x}{\verb{[vlightr_highlight]}

A highlighted vector to ID.}

\item{no_id}{\verb{[logical(1)]}

The identifier assigned to elements for which no condition returns \code{TRUE}.
Defaults to \code{NA}.}
}
\description{
\code{\link[=condition_id]{condition_id()}} identifies elements of a highlighted vector \code{x} by the
conditions (\code{attr(x, "conditions")}) that each element meets. Each element's
identifier is a comma separated string of descriptions
(\code{attr(x, "description")}) of the conditions met by the element.
}
\examples{
x <- highlight(1:5)

}
\seealso{
Other condition utilities: 
\code{\link{condition_split}()}
}
\concept{condition utilities}
