% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{highlight_functions}
\alias{highlight_functions}
\alias{highlight_functions<-}
\alias{tests}
\alias{tests<-}
\title{Get or set the test and formatter functions of a highlighted vector or
highlighter function}
\usage{
highlight_functions(x, as_list_of = c("list", "highlighter"))

highlight_functions(x) <- value

tests(x)

tests(x) <- value
}
\description{
Get or set the test and formatter functions of a highlighted vector or
highlighter function
}
