% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{re_highlight}
\alias{re_highlight}
\alias{rl}
\title{Restore a highlighted vector's formatting}
\usage{
re_highlight(x, ...)

rl(x, ...)
}
\arguments{
\item{x}{\verb{[vector]}

A vector to highlight with the condition formatting of highlights supplied to
\code{...}. If \code{x} is already a highlighted vector, then it's conditional
formatting is combined with that of the highlights supplied to \code{...}.}

\item{...}{\verb{[vlighter_highlight]}

A highlighted vector used to highlight \code{x}.}
}
\description{
\code{re_highlight()} adds the conditional formatting of highlighted vectors in
\code{...} to a highlighted or un-highlighted vector \code{x}.

The primary function of \code{re_highlight()} is to restore the format method of a
highlighted vector after un-highlighting it during an intermediate operation.
See the \emph{Highlight, Un-Highlight, Re-Highlight} section of
\code{vignette("vlightr")} for more details.

\code{re_highlight(un_highlight(x_hl), x_hl)} and \code{x_hl} are equivalent for a
highlighted vector \code{x_hl}.

\code{re_highlight()} and \code{rl()} are synonyms.
}
\examples{
# TODO

}
