% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight.R
\name{re_highlight}
\alias{re_highlight}
\alias{rl}
\title{Restore a highlighted vector's formatting}
\usage{
re_highlight(x, ...)

rl(x, ...)
}
\arguments{
\item{x}{\verb{[vector]}

A vector to highlight with the condition formatting of highlights supplied to
\code{...}. If \code{x} is already a highlighted vector, then it's conditional
formatting is combined with that of the highlights supplied to \code{...}.}

\item{...}{\verb{[vlighter_highlight]}

A highlighted vector used to highlight \code{x}. The \code{conditions}, \code{formatters},
\code{description}, \code{precedence}, \code{format_once}, \code{init_formatter}, and
\code{last_formatter} of each \code{...} argument are appended to that of \code{x}, if \code{x}
is a highlighted vector, and are supplied to \code{highlight(x, ...)} otherwise.

The \code{format_once}, \code{init_formatter}, and \code{last_formatter} attributes are
scalar, meaning they cannot be appended to.

The \code{format_once} attribute of the output vector is set to \code{TRUE} if any of
\code{x} or \code{...} have \code{format_once} is \code{TRUE}. Otherwise, \code{format_once} is set to
\code{FALSE}.

\code{init_formatter} and \code{last_formatter} are taken from \code{x}, if they are defined
(i.e. not \code{NULL}) for \code{x}, and are otherwise taken from the first argument in \code{...}
with a defined \code{init_formatter} and \code{last_formatter} respectively.}
}
\description{
\code{re_highlight()} adds the conditional formatting of highlighted vectors in
\code{...} to a highlighted or un-highlighted vector \code{x}.

The primary function of \code{re_highlight()} is to restore the format method of a
highlighted vector after un-highlighting it during an intermediate operation.
See the \strong{Highlight, Un-Highlight, Re-Highlight} section of
\code{vignette("vlightr")} for more details.

\code{re_highlight(un_highlight(x_hl), x_hl)} and \code{x_hl} are equivalent for a
highlighted vector \code{x_hl}.

\code{re_highlight()} and \code{rl()} are synonyms.
}
