% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{stylers}
\alias{stylers}
\alias{background}
\alias{bg}
\alias{color}
\alias{colour}
\alias{color_rep}
\alias{colour_rep}
\alias{background_rep}
\alias{style}
\title{Generate a formatter function to color text}
\usage{
background(x)

bg(x)

color(x)

colour(x)

color_rep(...)

colour_rep(...)

background_rep(...)

style(x)
}
\arguments{
\item{x}{\verb{[character(1)]}

For \code{background()} and \code{color()}, the color that the output function applies
to the text background or text of a character vector, specified by a:
\itemize{
\item recognized R color name (see \code{\link[grDevices:colors]{grDevices::colors()}})
\item \code{cli} package color (e.g. \code{"yellow"}, \code{"br_blue"})
\item 6- or 8-digit hexadecimal color string (e.g. "#ff0000")
}

For \code{style}, the text decoration that the output function adds to the text of
a character vector, specified by a \code{cli} package style (e.g. \code{"bold"},
\code{"italic"}, \code{"underline"}).

See \code{\link[cli:ansi-styles]{cli::style_bold()}} and friends for recognized \code{cli} styles and colors.}

\item{...}{\verb{[character]}

For \code{background_rep()} and \code{color_rep()}, the alternating series of colors that
the output function applies to the background or text of a character vector.

All inputs to \code{...} are concatenated into one character vector.}
}
\value{
A function which returns a styled ANSI string vector.
}
\description{
\code{color()}, \code{color_rep()}, \code{background()}, and \code{background_rep()} generate
a function that colors the text or background of an input string. \code{style()}
generates a function that bolds, underlines, or otherwise styles an input
string. These function factories are useful for quickly creating new formatting
functions for use in \code{\link[=highlight]{highlight()}}, \code{\link[=highlighter]{highlighter()}}, and friends.

\code{color()} and \code{background()} color the input character vector with a single
color. \code{color_rep()} and \code{background_rep()} color the input vector with a
repeating pattern of colors.

\code{colour()} is a synonym for \code{color()} and \code{bg()} a synonym for \code{background()}.
}
\examples{
color_red <- color("red")
color_red(1:3)

background_yellow <- background("yellow")
background_yellow(LETTERS[1:3])

style_bold <- style("bold")
style_bold(c(TRUE, FALSE, NA))

halloween <- color_rep("orange", "black")
halloween("Trick or treat?")

waldo <- background_rep("red", "white")
waldo(c("I can't", "find", "him"))

}
\concept{styles}
