% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{stylers}
\alias{stylers}
\alias{background}
\alias{bg}
\alias{color}
\alias{colour}
\alias{style}
\alias{wrap}
\alias{label}
\title{Generate a formatter function to style text}
\usage{
background(x)

bg(x)

color(x)

colour(x)

style(x)

wrap(left = "[", right = "]")

label(string, left = "[", right = "]")
}
\arguments{
\item{x}{\verb{[character(1)]}

For \code{background()} and \code{color()}, the color that the output function applies
to the text background or text of a character vector, specified by a:
\itemize{
\item recognized R color name (see \code{\link[grDevices:colors]{grDevices::colors()}})
\item \code{cli} package color (e.g. \code{"yellow"}, \code{"br_blue"})
\item 6- or 8-digit hexadecimal color string (e.g. "#ff0000")
}

For \code{style}, the text effect that the output function adds to the text of a
character vector, specified by a \code{cli} package style (e.g. \code{"bold"},
\code{"italic"}, \code{"underline"}).

See \code{\link[cli:ansi-styles]{cli::style_bold()}} and friends for recognized \code{cli} styles and colors.}

\item{left, right}{\verb{[character(1)]}

For \code{wrap()}, a prefix (\code{left}) or suffix \code{right} that the output function
pastes onto a character vector. By default, \code{left} is \code{"["} and \code{right}
is \code{"]"}.}
}
\value{
A function which returns a styled character or ANSI string vector.
}
\description{
Function factors used to generate new formatting functions for use in the
\code{formatters}, \code{init_formatter}, and \code{last_formatter} arguments of highlight
functions (e.g. \code{\link[=highlight]{highlight()}}, \code{\link[=highlighter]{highlighter()}}).
}
\examples{
color_red <- color("red")
color_red(1:3)

background_yellow <- background("yellow")
background_yellow(LETTERS[1:3])

style_bold <- style("bold")
style_bold(c(TRUE, FALSE, NA))

embrace <- wrap("(", ")")
embrace(c(2.2, 3.3, 4.4))
}
\concept{styles}
